requirements.txt
gspread==6.2.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3
numpy==2.0.2
openpyxl==3.1.5
pandas==2.3.3
pymongo==4.15.4
python-dotenv==1.2.1
requests==2.32.5
streamlit==1.50.0
supabase==2.24.0
tornado==6.5.2
certifi==2024.8.30
